refactor(direct-control): extract notification view source
Move the embedded App UI readPlayNotifications source into packages/civ7-direct-control/src/play/notifications/view.ts and import it from the package facade. Keep the public wrapper and build-command composition in index.ts for the later composition pass.

Record the source-owner status in the support modularization OpenSpec workstream. This is local package source relocation proof only, not runtime or live-game proof.

Validation: bun run --cwd packages/civ7-direct-control test -- play-notification-view.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
